font and lib move
