+++ src/app/components/Footer/index.jsx
@@ -5,7 +5,7 @@ import {
   C_WHITE,
   FF_NEWS_SANS_REG,
   GEL_SPACING_DBL,
-} from '../../../lib/constants/styles';
+} from '../../lib/constants/styles';
 
 const StyledFooter = styled.footer`
   background-color: ${C_STORM};
