+++ src/app/components/Headline/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { extractText } from '../../helpers/blockHandlers';
 import { textPropTypes, textDefaultPropTypes } from '../../models/proptypes';
-import { C_EBON, FF_NEWS_SANS_REG } from '../../../lib/constants/styles';
+import { C_EBON, FF_NEWS_SANS_REG } from '../../lib/constants/styles';
 
 const StyledHeadline = styled.h1`
   color: ${C_EBON};
